Move helpers to an extension
