+++ .gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 /.build
 /Packages
+Package.resolved
 xcuserdata/
 DerivedData/
 .swiftpm/configuration/registries.json
